# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dango DEX Tools API
 * OpenAPI spec version: 0.1.0
 */
export type InitiatePaymentBodyAmount =
  (typeof InitiatePaymentBodyAmount)[keyof typeof InitiatePaymentBodyAmount];

export const InitiatePaymentBodyAmount = {
  NUMBER_50000: 50000,
  NUMBER_100000: 100000,
  NUMBER_150000: 150000,
} as const;

export interface InitiatePaymentBody {
  amount: InitiatePaymentBodyAmount;
  name?: string;
  email?: string;
}

export interface InitiatePaymentResponse {
  donationId: string;
  qrString: string;
  amountToPay: number;
  pgFee: number;
}

export type PollPaymentResponseStatus =
  (typeof PollPaymentResponseStatus)[keyof typeof PollPaymentResponseStatus];

export const PollPaymentResponseStatus = {
  PENDING: "PENDING",
  SUCCESS: "SUCCESS",
  FAILED: "FAILED",
} as const;

export interface PollPaymentResponse {
  status: PollPaymentResponseStatus;
  token?: string | null;
  expiresAt?: string | null;
  amount?: number | null;
}

export interface VerifyTokenBody {
  token: string;
}

export interface VerifyTokenResponse {
  valid: boolean;
  expiresAt?: string | null;
  amount?: number | null;
  reason?: string | null;
}

export interface HealthStatus {
  status: string;
}

export type RerangeMode = (typeof RerangeMode)[keyof typeof RerangeMode];

export const RerangeMode = {
  off: "off",
  conservative: "conservative",
  moderate: "moderate",
  aggressive: "aggressive",
} as const;

export type OrderType = (typeof OrderType)[keyof typeof OrderType];

export const OrderType = {
  long: "long",
  short: "short",
  neutral: "neutral",
} as const;

export type ExecutionMode = (typeof ExecutionMode)[keyof typeof ExecutionMode];

export const ExecutionMode = {
  aggressive: "aggressive",
  normal: "normal",
  passive: "passive",
} as const;

export interface GridBot {
  id: number;
  name: string;
  pair: string;
  orderType: OrderType;
  lowerPrice: number;
  upperPrice: number;
  gridCount: number;
  investmentAmount: number;
  executionMode: ExecutionMode;
  rerangeMode: RerangeMode;
  budgetStop?: number | null;
  isActive: boolean;
  totalPnl: number;
  rerangeCount: number;
  lastRerangeAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateGridBotBody {
  name: string;
  pair: string;
  orderType: OrderType;
  lowerPrice: number;
  upperPrice: number;
  gridCount: number;
  investmentAmount: number;
  executionMode: ExecutionMode;
  rerangeMode: RerangeMode;
  budgetStop?: number | null;
}

export interface UpdateGridBotBody {
  name?: string;
  pair?: string;
  orderType?: OrderType;
  lowerPrice?: number;
  upperPrice?: number;
  gridCount?: number;
  investmentAmount?: number;
  executionMode?: ExecutionMode;
  rerangeMode?: RerangeMode;
  budgetStop?: number | null;
  isActive?: boolean;
}

export interface BotLog {
  id: number;
  botId: number;
  eventType: string;
  message: string;
  priceAtEvent?: number | null;
  createdAt: string;
}

export interface MarketSummary {
  totalBots: number;
  activeBots: number;
  totalPnl: number;
  totalReranges: number;
  recentActivity: BotLog[];
}

export type BotStatsByMode = { [key: string]: number };

export interface BotStats {
  byMode: BotStatsByMode;
  avgPnl: number;
  topPerformer?: GridBot | null;
}

export interface MarketPrice {
  symbol: string;
  pair: string;
  price: number;
  change24h: number;
  high24h: number;
  low24h: number;
  volume24h: number;
  source: string;
}

export type ListBotLogsParams = {
  botId?: number;
  limit?: number;
};