# lib/api-zod/src/generated/types/botStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dango DEX Tools API
 * OpenAPI spec version: 0.1.0
 */
import type { BotStatsByMode } from "./botStatsByMode";
import type { GridBot } from "./gridBot";

export interface BotStats {
  byMode: BotStatsByMode;
  avgPnl: number;
  topPerformer?: GridBot | null;
}