# lib/api-zod/src/generated/types/initiatePaymentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dango DEX Tools API
 * OpenAPI spec version: 0.1.0
 */
import type { InitiatePaymentBodyAmount } from "./initiatePaymentBodyAmount";

export interface InitiatePaymentBody {
  amount: InitiatePaymentBodyAmount;
  name?: string;
  email?: string;
}