# lib/api-zod/src/generated/types/initiatePaymentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dango DEX Tools API
 * OpenAPI spec version: 0.1.0
 */

export interface InitiatePaymentResponse {
  donationId: string;
  qrString: string;
  amountToPay: number;
  pgFee: number;
}