# lib/api-zod/src/generated/types/marketSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dango DEX Tools API
 * OpenAPI spec version: 0.1.0
 */
import type { BotLog } from "./botLog";

export interface MarketSummary {
  totalBots: number;
  activeBots: number;
  totalPnl: number;
  totalReranges: number;
  recentActivity: BotLog[];
}