# lib/api-zod/src/generated/types/pollPaymentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dango DEX Tools API
 * OpenAPI spec version: 0.1.0
 */
import type { PollPaymentResponseStatus } from "./pollPaymentResponseStatus";

export interface PollPaymentResponse {
  status: PollPaymentResponseStatus;
  token?: string | null;
  expiresAt?: string | null;
  amount?: number | null;
}