# lib/api-zod/src/generated/types/updateGridBotBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dango DEX Tools API
 * OpenAPI spec version: 0.1.0
 */
import type { ExecutionMode } from "./executionMode";
import type { OrderType } from "./orderType";
import type { RerangeMode } from "./rerangeMode";

export interface UpdateGridBotBody {
  name?: string;
  pair?: string;
  orderType?: OrderType;
  lowerPrice?: number;
  upperPrice?: number;
  gridCount?: number;
  investmentAmount?: number;
  executionMode?: ExecutionMode;
  rerangeMode?: RerangeMode;
  budgetStop?: number | null;
  isActive?: boolean;
}