# lib/api-zod/src/generated/types/verifyTokenResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dango DEX Tools API
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyTokenResponse {
  valid: boolean;
  expiresAt?: string | null;
  amount?: number | null;
  reason?: string | null;
}